tech(api): migrate create target profile to its bounded context
