+++ api/src/prescription/target-profile/application/http-error-mapper-configuration.js
@@ -1,0 +1,13 @@
+import { HttpErrors } from '../../../shared/application/http-errors.js';
+import { TargetProfileCannotBeCreated } from '../domain/errors.js';
+
+const targetProfileDomainErrorMappingConfiguration = [
+  {
+    name: TargetProfileCannotBeCreated.name,
+    httpErrorFn: (error) => {
+      return new HttpErrors.UnprocessableEntityError(error.message);
+    },
+  },
+];
+
+export { targetProfileDomainErrorMappingConfiguration };
