+++ api/src/prescription/target-profile/domain/constants.js
@@ -1,0 +1,3 @@
+const TARGET_PROFILE_COPY_NAME_PREFIX = '[Copie] ';
+
+export { TARGET_PROFILE_COPY_NAME_PREFIX };
