+++ api/src/prescription/target-profile/domain/errors.js
@@ -1,0 +1,9 @@
+import { DomainError } from '../../../shared/domain/errors.js';
+
+class TargetProfileCannotBeCreated extends DomainError {
+  constructor(message = 'Erreur lors de la création du profil cible.') {
+    super(message);
+  }
+}
+
+export { TargetProfileCannotBeCreated };
