+++ api/src/prescription/target-profile/domain/usecases/create-target-profile.js
@@ -1,5 +1,5 @@
-import { TargetProfileForCreation } from '../../../src/prescription/target-profile/domain/models/TargetProfileForCreation.js';
-import { TargetProfileCannotBeCreated } from '../../../src/shared/domain/errors.js';
+import { TargetProfileCannotBeCreated } from '../errors.js';
+import { TargetProfileForCreation } from '../models/TargetProfileForCreation.js';
 
 const createTargetProfile = async function ({
   targetProfileCreationCommand,
