+++ api/src/prescription/target-profile/domain/validators/creation-command-validation.js
@@ -1,6 +1,6 @@
 import Joi from 'joi';
 
-import { EntityValidationError } from '../../../../src/shared/domain/errors.js';
+import { EntityValidationError } from '../../../../shared/domain/errors.js';
 import { schema as base } from './base-validation.js';
 
 const schema = base.keys({
