+++ api/src/shared/domain/constants.js
@@ -60,8 +60,6 @@ const VALIDATION_ERRORS = {
   UNICITY_COLUMNS_REQUIRED: 'UNICITY_COLUMNS_REQUIRED',
 };
 
-const TARGET_PROFILE_COPY_NAME_PREFIX = '[Copie] ';
-
 const PIX_COUNT_BY_LEVEL = 8;
 const COMPETENCES_COUNT = 16;
 const MAX_REACHABLE_PIX_BY_COMPETENCE = config.features.maxReachableLevel * PIX_COUNT_BY_LEVEL;
@@ -196,7 +194,6 @@ export {
   PIX_ORIGIN,
   STUDENT_RECONCILIATION_ERRORS,
   SUPPORTED_LOCALES,
-  TARGET_PROFILE_COPY_NAME_PREFIX,
   UNCERTIFIED_LEVEL,
   V3_REPRODUCIBILITY_RATE,
   VALIDATION_ERRORS,
