+++ api/tests/prescription/target-profile/unit/domain/usecases/create-target-profile_test.js
@@ -1,7 +1,7 @@
-import { createTargetProfile } from '../../../../lib/domain/usecases/create-target-profile.js';
-import { TargetProfileCannotBeCreated } from '../../../../src/shared/domain/errors.js';
-import { categories } from '../../../../src/shared/domain/models/TargetProfile.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { TargetProfileCannotBeCreated } from '../../../../../../src/prescription/target-profile/domain/errors.js';
+import { createTargetProfile } from '../../../../../../src/prescription/target-profile/domain/usecases/create-target-profile.js';
+import { categories } from '../../../../../../src/shared/domain/models/TargetProfile.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | create-target-profile', function () {
   let targetProfileAdministrationRepositoryStub;
