+++ api/tests/prescription/target-profile/unit/domain/validators/creation-command-validation_test.js
@@ -1,7 +1,7 @@
-import * as creationCommandValidator from '../../../../../lib/domain/validators/target-profile/creation-command-validation.js';
-import { EntityValidationError } from '../../../../../src/shared/domain/errors.js';
-import { categories } from '../../../../../src/shared/domain/models/TargetProfile.js';
-import { expect } from '../../../../test-helper.js';
+import * as creationCommandValidator from '../../../../../../src/prescription/target-profile/domain/validators/creation-command-validation.js';
+import { EntityValidationError } from '../../../../../../src/shared/domain/errors.js';
+import { categories } from '../../../../../../src/shared/domain/models/TargetProfile.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Validators | target-profile/creationCommandValidator', function () {
   describe('#validate', function () {
